% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP_APITopLevelFunctions_History.R
\name{checkAPIResponse}
\alias{checkAPIResponse}
\title{checkAPIResponse}
\usage{
checkAPIResponse(statusCode)
}
\arguments{
\item{statusCode}{The status code contained in the endpoint response.}
}
\value{
It throws an ERROR if the status code corresonds to an error code.
It prints a message if the status code corresponds to a successfull code.
}
\description{
Internal function to check the status of responses from endpoints.
}
\author{
Maria Gomez
}
