% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP_APITopLevelFunctions_History.R
\name{downloadAnalyticsFile}
\alias{downloadAnalyticsFile}
\title{downloadAnalyticsFile}
\usage{
downloadAnalyticsFile(url, analyticsFileName, outFolder)
}
\arguments{
\item{url}{Url of a RavenPack endpoint.}

\item{analyticsFileName}{Name of the file to download from the endpoint.}

\item{outFolder}{The local folder to store the downloaded files.}
}
\value{
TRUE if the operation succeeds. An ERROR otherwise.
}
\description{
Internal function to download files from a RavenPack end-point.
}
\author{
Maria Gomez
}
