% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP_SelfServiceAPI_EndPoints.R
\name{RP_APICancelRequest}
\alias{RP_APICancelRequest}
\title{RP_APICancelRequest}
\usage{
RP_APICancelRequest(APIHandler, token)
}
\arguments{
\item{APIHandler}{An API handler, created using RP_CreateAPIHandler.}

\item{token}{A string that identifies the request to be cancelled.}
}
\value{
An informative message.
}
\description{
Cancels a Request. Only "Enqueued" queries can be cancelled.
Please note that if the job has finished, a 404 response will be returned as the job token is no longer valid.
}
\author{
Jose A. Guerrero-Colon
}
