% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP_SelfServiceAPI_EndPoints.R
\name{RP_APICheckFileAvailability}
\alias{RP_APICheckFileAvailability}
\title{RP_APICheckFileAvailability}
\usage{
RP_APICheckFileAvailability(APIHandler, token)
}
\arguments{
\item{APIHandler}{An API handler, created using RP_CreateAPIHandler.}

\item{token}{A string that identifies the request.}
}
\value{
Information about the status of the request.
\itemize{
Response parameters:
\item{status: } {One of the following: [enqueued, processing, complete]}
\item{size: } {File size.}
\item{url: } {Link to download the file.}
\item{checksum: } {Checksum code.}}
}
\description{
Checks the availability of the datafile. After submitting a request to generate a datafile using the datafile endpoint, a token is returned which may be supplied to this endpoint in order to find the status of the datafile generation job.
When the job is complete, the status will be updated to completed and the size and URL will be set.
}
\author{
Jose A. Guerrero-Colon
}
