% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP_SelfServiceAPI_EndPoints.R
\name{RP_APICreateDataSet}
\alias{RP_APICreateDataSet}
\title{RP_APICreateDataSet}
\usage{
RP_APICreateDataSet(APIHandler, payload)
}
\arguments{
\item{APIHandler}{An API handler, created using RP_CreateAPIHandler}

\item{payload}{A string in JSON format with the request body parameters. See the RavenPack Analytics User Guide for more details on the fields in RavenPack Analytics data.}
}
\value{
A dataset_uuid identifying the created dataset.
}
\description{
Creates a dataset. This endpoint allows you to create a new dataset definition on ravenpack.com.
A dataset allows you to query a subset of RavenPack data by specifying filters and fields available in RavenPack Analytics.
}
\author{
Jose A. Guerrero-Colon
}
