% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP_SelfServiceAPI_EndPoints.R
\name{RP_APIDeleteDataSet}
\alias{RP_APIDeleteDataSet}
\title{RP_APIDeleteDataSet}
\usage{
RP_APIDeleteDataSet(APIHandler, datasetUUID)
}
\arguments{
\item{APIHandler}{An API handler, created using RP_CreateAPIHandler}

\item{datasetUUID}{A string with the dataset identifier to delete.}
}
\value{
A list with the delete results if the dataset was successfully deleted. An Error otherwise.
The delete results contains a 'message' field.
}
\description{
Deletes a dataset.
}
\author{
Jose A. Guerrero-Colon
}
