% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP_APITopLevelFunctions.R
\name{RP_APIDownloadFile}
\alias{RP_APIDownloadFile}
\title{RP_APIDownloadFile}
\usage{
RP_APIDownloadFile(APIHandler, statusInfo, outputFile)
}
\arguments{
\item{APIHandler}{An API handler, created using RP_CreateAPIHandler.}

\item{statusInfo}{Request status coming from RP_APICheckFileAvailability().}

\item{outputFile}{Output filename.}
}
\value{
TRUE if the download was successful. FALSE otherwise.
}
\description{
Downloads a request file
}
\author{
Jose A. Guerrero-Colon
}
