% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP_APITopLevelFunctions_History.R
\name{RP_APIDownloadFullHistory}
\alias{RP_APIDownloadFullHistory}
\title{RP_APIDownloadFullHistory}
\usage{
RP_APIDownloadFullHistory(APIHandler, outFolderPath = "~/Downloads",
  onlyCompany = TRUE)
}
\arguments{
\item{APIHandler}{An API handler, created using RP_CreateAPIHandler.}

\item{outFolderPath}{The local folder to store the downloaded archive files.}

\item{onlyCompany}{TRUE as default. Whether to download only analytics for companies of for all type of entities.}
}
\value{
TRUE if the operation succeeds. An ERROR otherwise.
As output a set of zip files will be downloaded in the output folder indicated as parameter.
}
\description{
This function downloads the full archive of RavenPack Analytics.
This function is normally used for bulk loading the archive into a database.
Files downloaded are yearly zip files containing monthly CSV files, up to the end of the prior month, relative to today.
}
\author{
Maria Gomez
}
