% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP_SelfServiceAPI_EndPoints.R
\name{RP_APIGetDataFileCount}
\alias{RP_APIGetDataFileCount}
\title{RP_APIGetDataFileCount}
\usage{
RP_APIGetDataFileCount(APIHandler, payload, datasetUUID)
}
\arguments{
\item{APIHandler}{An API handler, created using RP_CreateAPIHandler.}

\item{payload}{A string in JSON format with the request body parameters.}

\item{datasetUUID}{A string with the dataset identifier.}
}
\value{
The count.
}
\description{
Gets the count of the dataset for a time range described on payload. The datafile count endpoint allows you to find out how many rows a particular datafile will contain before actually generating it.
In general it is a good idea to use this in order to determine if a particular datafile will be too large and will need to be broken up into smaller subsets.
Please note: When requesting a count for a dataset defined with granular frequency, the count will be an exact count of the records that match in the time-range specified. When requesting a count for a dataset defined with daily frequency, the count is an estimate and has a margin of error typically less then 1%, varying with the size of the dataset.
}
\author{
Jose A. Guerrero-Colon
}
