% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP_SelfServiceAPI_EndPoints.R
\name{RP_APIGetDataSet}
\alias{RP_APIGetDataSet}
\title{RP_APIGetDataSet}
\usage{
RP_APIGetDataSet(APIHandler, datasetUUID)
}
\arguments{
\item{APIHandler}{An API handler, created using RP_CreateAPIHandler}

\item{datasetUUID}{A string with the dataset identifier.}
}
\value{
The dataset information in JSON format.
}
\description{
Retrieves a dataset information. Provides the full specification for a single dataset
}
\author{
Jose A. Guerrero-Colon
}
