% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP_SelfServiceAPI_EndPoints.R
\name{RP_APIGetDataSetJSON}
\alias{RP_APIGetDataSetJSON}
\title{RP_APIGetDataSetJSON}
\usage{
RP_APIGetDataSetJSON(APIHandler, payload, datasetUUID)
}
\arguments{
\item{APIHandler}{An API handler, created using RP_CreateAPIHandler.}

\item{payload}{A string in JSON format with the request body parameters.}

\item{datasetUUID}{A string with the dataset identifier.}
}
\value{
A data.table with the requested dataset.
}
\description{
The JSON dataset endpoint allows data to be requested synchronously and in JSON format. A predefined dataset must be supplied and the fields property may be overriden.
Please note: Currently this endpoint only supports datasets with “daily” frequency.
}
\author{
Jose A. Guerrero-Colon
}
