% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP_SelfServiceAPI_EndPoints.R
\name{RP_APIGetDataSetPreview}
\alias{RP_APIGetDataSetPreview}
\title{RP_APIGetDataSetPreview}
\usage{
RP_APIGetDataSetPreview(APIHandler, payload, datasetUUID)
}
\arguments{
\item{APIHandler}{An API handler, created using RP_CreateAPIHandler.}

\item{payload}{A string in JSON format with the request body parameters.}

\item{datasetUUID}{A string with the dataset identifier to preview.}
}
\value{
A data.table with the preview.
}
\description{
Gets a preview from a dataset. Please note: Currently this endpoint only supports datasets with “daily” frequency. The sample will contain up to five entities from the entities supported by the dataset, and a maximum of 10 days per entities.
In addition, a rollup for the whole dataset is never returned in the preview.
}
\author{
Jose A. Guerrero-Colon
}
