% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndPoints_Document.R
\name{RP_APIGetDocumentProperties}
\alias{RP_APIGetDocumentProperties}
\title{RP_APIGetDocumentProperties}
\usage{
RP_APIGetDocumentProperties(APIHandler, rpStoryId)
}
\arguments{
\item{APIHandler}{An API handler, created using RP_CreateAPIHandler}

\item{rpStoryId}{A RavenPack identifier for a story (i.e. rp_story_id)}
}
\value{
It returns a json containing the URL to access the content of a story.
}
\description{
This function retrieves a URL for accessing the content of a story.
In the case of a story from MoreOver provider, the original_url will be returned.
In the case of premium providers, a URL to the RavenPack secure platform will be returned.
}
\author{
Maria Gomez
}
