% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP_SelfServiceAPI_EndPoints.R
\name{RP_APIGetFullAdhocJSON}
\alias{RP_APIGetFullAdhocJSON}
\title{RP_APIGetFullAdhocJSON}
\usage{
RP_APIGetFullAdhocJSON(APIHandler, payload)
}
\arguments{
\item{APIHandler}{An API handler, created using RP_CreateAPIHandler.}

\item{payload}{A string in JSON format with the request body parameters.}

\item{datasetUUID}{A string with the dataset identifier to preview.}
}
\value{
A data.table with the preview.
}
\description{
The JSON dataset endpoint allows data to be requested synchronously in JSON format and without having previously defined a dataset.
The endpoint provides similar parameters to the ones used when creating a dataset so that RavenPack Analytics data can be queried in a more adhoc way.
Please note: Currently this endpoint only supports datasets with “daily” frequency.
}
\details{
Limitations: will show the top 500 entities (in terms of volume), max lookback window will be 366 days, and max time range will be 366 days.
In the case of a timeout, the response will be a generic bad-request HTTP 400 code specifying that the requested data is too large and to split the request into smaller chunks.
}
\author{
Jose A. Guerrero-Colon
}
