% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP_SelfServiceAPI_EndPoints.R
\name{RP_APIGetReferenceData}
\alias{RP_APIGetReferenceData}
\title{RP_APIGetReferenceData}
\usage{
RP_APIGetReferenceData(APIHandler, params = list())
}
\arguments{
\item{APIHandler}{An API handler, created using RP_CreateAPIHandler}

\item{params}{A list with the request query parameters.
\itemize{\item{entity_type: }{Optional. A single entity type. If not provided all are requested.}}}
}
\value{
An url with a link to the csv file upon success. An http server response otherwise.
}
\description{
Retrieve an entity reference data file. An entity mapping file contains information pertaining to entities in RavenPack’s entity universe.
Files are updated on a daily basis after midnight UTC. It is possible to request a reference data file for a specific entity type by supplying the entity_type parameter.
For more information, please see the Reference Service section in the RavenPack Analytics User Guide.
}
\author{
Jose A. Guerrero-Colon
}
