% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP_SelfServiceAPI_EndPoints.R
\name{RP_APIListDataSet}
\alias{RP_APIListDataSet}
\title{RP_APIListDataSet}
\usage{
RP_APIListDataSet(APIHandler, params = list())
}
\arguments{
\item{APIHandler}{An API handler, created using RP_CreateAPIHandler}

\item{params}{A list with the request query parameters.
\itemize{ \item{tags: }{Optional. A list of tags that should be used to filter the list of datasets returned.}
\item{scope: }{Optional. A list with the desired scope: [public, shared, private (default)]}}}
}
\value{
A data.table with the datasets.
}
\description{
List all available datasets. Get a list of the datasets that you have permission to access. You may filter the list by tags and search for only the list of datasets that have the tags specified, or you may filter by scope, and return only the datasets that are Public to everyone, Shared with you by someone else using RavenPack or Private datasets that were created by you.
The list of datasets returns the dataset_uuid and name for each dataset. It is possible to request the full set of information that defines a dataset using the API.
}
\author{
Jose A. Guerrero-Colon
}
