% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP_SelfServiceAPI_EndPoints.R
\name{RP_APIMappingRequest}
\alias{RP_APIMappingRequest}
\title{RP_APIMappingRequest}
\usage{
RP_APIMappingRequest(APIHandler, payload)
}
\arguments{
\item{APIHandler}{An API handler, created using RP_CreateAPIHandler}

\item{payload}{A string in JSON format with the request body parameters.}
}
\value{
A list with the mapping information, upon success. Error information otherwise.
}
\description{
The entity-mapping endpoint may be used to map from a universe of entity or security identifiers into RavenPack’s entity universe.
One may pass in identifiers such as entity names, listings, ISIN values, CUSIP values, etc. and the endpoint will return the corresponding RP_ENTITY_ID for the possible matches.
In the event that the entity mapping API is unable to match the requested entity to an entity in the RavenPack entity universe, there will be no mapped entities and the requested data is returned as an error.
In the event that multiple entities are matched, the entities will be returned ranked with a relative score, which may be used to automatically filter or sort for further analysis.
}
\author{
Jose A. Guerrero-Colon
}
