% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP_SelfServiceAPI_EndPoints.R
\name{RP_APIModifyDataSet}
\alias{RP_APIModifyDataSet}
\title{RP_APIModifyDataSet}
\usage{
RP_APIModifyDataSet(APIHandler, payload, datasetUUID)
}
\arguments{
\item{APIHandler}{An API handler, created using RP_CreateAPIHandler.}

\item{payload}{A string in JSON format with the request body parameters.}

\item{datasetUUID}{A string with the dataset identifier to modify.}
}
\value{
TRUE if the dataset was successfully deleted. An error message otherwise.
}
\description{
Modify an existing dataset on ravenpack.com. When modifying a dataset, it is possible to provide just the parameters that you wish to modify.
Any parameters that are not included in the request will retain their value and will not be modified.
}
\author{
Jose A. Guerrero-Colon
}
