% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP_SelfServiceAPI_EndPoints.R
\name{RP_APIRequestDataFile}
\alias{RP_APIRequestDataFile}
\title{RP_APIRequestDataFile}
\usage{
RP_APIRequestDataFile(APIHandler, payload, datasetUUID)
}
\arguments{
\item{APIHandler}{An API handler, created using RP_CreateAPIHandler.}

\item{payload}{A string in JSON format with the request body parameters.}

\item{datasetUUID}{A string with the dataset identifier.}
}
\value{
A list with the request token and the expected time.
}
\description{
The datafile endpoint allows a file to be generated for a particular dataset for any period from the year 2000 to present.
Data may be retrieved in CSV or Excel format and may be compressed for transmission via HTTP.
Note: There is a limitation of 50 million records for generating granular data and 10 million records for aggregated data.
}
\author{
Jose A. Guerrero-Colon
}
