% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP_SelfServiceAPI_EndPoints.R
\name{RP_APIStatus}
\alias{RP_APIStatus}
\title{RP_APIStatus}
\usage{
RP_APIStatus(APIHandler)
}
\arguments{
\item{APIHandler}{An API handler, created using RP_CreateAPIHandler}
}
\value{
The status of the API ("OK" expected)
}
\description{
Request the status of the server. This endpoint allows you to check that your connection with the server is running correctly and that the server is functioning correctly.
}
\author{
Jose A. Guerrero-Colon
}
