% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP_SelfServiceAPI_EndPoints.R
\name{RP_APITaxonomy}
\alias{RP_APITaxonomy}
\title{RP_APITaxonomy}
\usage{
RP_APITaxonomy(APIHandler, payload)
}
\arguments{
\item{APIHandler}{An API handler, created using RP_CreateAPIHandler.}

\item{payload}{A string in JSON format with the request body parameters.}
}
\value{
A list with the taxonomy information (topic, group, type,...)
}
\description{
The RavenPack taxonomy is a comprehensive structure for httr::content classification.
It provides a definitive system categorizing structured and unstructured information, enabling analysis on thousands of entities including companies, products, people, organizations, places, and more.
 This endpoint allows you to query and browse the RavenPack taxonomy.
}
\author{
Jose A. Guerrero-Colon
}
