% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP_APITopLevelFunctions.R
\name{RP_APIWaitForJobCompletion}
\alias{RP_APIWaitForJobCompletion}
\title{RP_APIWaitForJobCompletion}
\usage{
RP_APIWaitForJobCompletion(APIHandler, token, timeout = 60)
}
\arguments{
\item{APIHandler}{An API handler, created using RP_CreateAPIHandler.}

\item{token}{A string that identifies the request. Check RP_APIRequestDataFile to generate a request.}

\item{timeout}{Time (in seconds) to wait.}
}
\value{
The final status of the job. A list containg two values:
         STATUS (String defining the final status: 'completed', 'error' or 'timeout'), and
         STATUSINFO status object coming from RP_APICheckFileAvailability().
}
\description{
Waits until the request file is available to download
}
\author{
Maria Gomez
}
