% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP_APITopLevelFunctions.R
\name{RP_APIDownloadFileWhenReady}
\alias{RP_APIDownloadFileWhenReady}
\title{RP_APIDownloadFileWhenReady}
\usage{
RP_APIDownloadFileWhenReady(APIHandler, token, outputFile, timeout = 60)
}
\arguments{
\item{APIHandler}{An API handler, created using RP_CreateAPIHandler.}

\item{token}{A string that identifies the request. Check RP_APIRequestDataFile to generate a request.}

\item{outputFile}{Output filename.}

\item{timeout:}{time (in seconds) to wait for download. Default 60.}
}
\value{
TRUE if the download was successful. An Error otherwise.
}
\description{
Downloads a request file when it is ready.
}
\author{
Maria Gomez
}
