% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP_SelfServiceAPI_EndPoints.R
\name{RP_CreateAPIHandler}
\alias{RP_CreateAPIHandler}
\title{RP_CreateAPIHandler}
\usage{
RP_CreateAPIHandler(APIKey)
}
\arguments{
\item{APIKey}{A valid API Key.}
}
\value{
An API Handler.
}
\description{
This functions creates an API Handler, needed to call all API functions.
}
\author{
Jose A. Guerrero-Colon
}
